Feature/fixgame

See merge request cpp-2025/strategy-game-1!6

+++ src/entities/DefaultTower.hpp
@@ -7,7 +7,7 @@ class Player;
 
 class DefaultTower : public Tower {
 public:
-  DefaultTower(int x, int y, bool isKingTower, Player* owner);
+  DefaultTower(float x, float y, bool isKingTower, Player* owner);
   void update(float deltaTime) override;
 };
